animation

+++ src/main/java/sep/fimball/model/blueprint/base/BaseElementJson.java
@@ -166,7 +166,7 @@ public class BaseElementJson
             /**
              * Die Genaugigkeit der Rotation beim Setzen und Bearbeiten im Editor, d.h. um wie viel Grad man Elemente von diesem Typ drehen kann.
              */
-            public double rotationAccuracy;
+            public int rotationAccuracy;
         }
 
         /**
